🚧 VEZQYS task: split Obsidian context tests
